# imported/Appgard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gard Security API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rounds:
    get:
      operationId: listRounds
      tags: [rounds]
      summary: List rounds
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, completed]
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of rounds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Round"
    post:
      operationId: createRound
      tags: [rounds]
      summary: Create a new round (supervisor/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoundBody"
      responses:
        "201":
          description: Round created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
  /rounds/{id}:
    get:
      operationId: getRound
      tags: [rounds]
      summary: Get round details with checkpoints
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Round with checkpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundDetail"
        "404":
          description: Not found
    patch:
      operationId: updateRound
      tags: [rounds]
      summary: Update round
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoundBody"
      responses:
        "200":
          description: Round updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
    delete:
      operationId: deleteRound
      tags: [rounds]
      summary: Delete round
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rounds/{id}/start:
    post:
      operationId: startRound
      tags: [rounds]
      summary: Start a round (guard)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Round started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
  /rounds/{id}/complete:
    post:
      operationId: completeRound
      tags: [rounds]
      summary: Complete a round
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Round completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List chat messages
      parameters:
        - name: channelId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /panic:
    get:
      operationId: listPanicAlerts
      tags: [panic]
      summary: List panic alerts
      responses:
        "200":
          description: List of panic alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PanicAlert"
    post:
      operationId: triggerPanic
      tags: [panic]
      summary: Trigger panic alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerPanicBody"
      responses:
        "201":
          description: Panic alert triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanicAlert"
  /panic/{id}/resolve:
    post:
      operationId: resolvePanic
      tags: [panic]
      summary: Resolve a panic alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanicAlert"
...
components:
  schemas:
...
    Round:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, completed]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        createdById:
          type: integer
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        totalCheckpoints:
          type: integer
        completedCheckpoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - companyName
        - status
        - createdById
        - totalCheckpoints
        - completedCheckpoints
        - createdAt
    RoundDetail:
      type: object
      properties:
        round:
          $ref: "#/components/schemas/Round"
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/Checkpoint"
      required:
        - round
        - checkpoints
    CreateRoundBody:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        description:
          type: string
        assignedToId:
          type: ["integer", "null"]
        scheduledAt:
          type: string
          format: date-time
      required:
        - name
        - companyName
    UpdateRoundBody:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        description:
          type: string
        assignedToId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, active, completed]
    Checkpoint:
      type: object
      properties:
        id:
          type: integer
        roundId:
          type: integer
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        orderIndex:
          type: integer
        checkedIn:
          type: boolean
        checkedInAt:
          type: ["string", "null"]
          format: date-time
        checkedInById:
          type: ["integer", "null"]
      required:
        - id
        - roundId
        - name
        - latitude
        - longitude
        - radius
        - orderIndex
        - checkedIn
...
    Message:
      type: object
      properties:
        id:
          type: integer
        channelId:
          type: string
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - channelId
        - senderId
        - senderName
        - content
        - createdAt
    SendMessageBody:
      type: object
      properties:
        channelId:
          type: string
        content:
          type: string
      required:
        - channelId
        - content
    PanicAlert:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [active, resolved]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - userId
        - userName
        - latitude
        - longitude
        - status
        - createdAt
    TriggerPanicBody:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        message:
          type: string
      required:
        - latitude
        - longitude
...