# imported/Appgard/lib/api-zod/src/generated/types/round.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gard Security API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoundStatus } from "./roundStatus";

export interface Round {
  id: number;
  name: string;
  companyName: string;
  /** @nullable */
  description?: string | null;
  status: RoundStatus;
  /** @nullable */
  assignedToId?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  createdById: number;
  /** @nullable */
  scheduledAt?: Date | null;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  completedAt?: Date | null;
  totalCheckpoints: number;
  completedCheckpoints: number;
  createdAt: Date;
}